windows: Skip some nbdkit-cow-filter tests
It's unclear exactly why, but some (but not all) Windows tests of the
cow filter fail with a strange error where we are unable to truncate
the temporary overlay file.  Document the error and skip these tests
for now.

(cherry picked from commit 88bfd25c1ac42701a02f7ec526867008016dafdd)

+++ tests/test-cow-on-read.sh
@@ -37,6 +37,13 @@ set -x
 requires_filter cow
 requires_nbdsh_uri
 
+# On Windows, calling ftruncate in the cow filter fails with:
+# nbdkit: memory[1]: error: ftruncate: File too large
+if is_windows; then
+   echo "$0: the cow filter needs to be fixed to work on Windows"
+   exit 77
+fi
+
 sock=$(mktemp -u /tmp/nbdkit-test-sock.XXXXXX)
 files="$sock cow-on-read.pid"
 rm -f $files
